Fix checkstyles issues.
