+++ src/main/java/org/fibsters/ComputeJob.java
@@ -1,0 +1,8 @@
+package org.fibsters;
+
+interface ComputeJob extends Runnable {
+    void setInputPayload(InputPayload inputPayload);
+
+    void getInputPayload();
+
+}
