+++ src/main/java/org/fibsters/ComputeJobPool.java
@@ -1,0 +1,16 @@
+package org.fibsters;
+
+interface ComputeJobPool {
+    // i.e. public class MyComputeJobPool implements ComputeJobPool { ... } blah blah
+    void addJob(ComputeJob job);
+
+    void removeJob(ComputeJob job);
+
+    void processJob(ComputeJob job);
+
+    void processAllJobs();
+
+    void getJobStatus(ComputeJob job);
+
+    void getJobResults(ComputeJob job);
+}
