+++ src/main/java/org/fibsters/ComputeJobStatus.java
@@ -1,0 +1,9 @@
+package org.fibsters;
+
+interface ComputeJobStatus {
+    public String getStatus();
+
+    public String getErrorMessage();
+
+    public String getErrorType();
+}
