+++ src/main/java/org/fibsters/DataStorage.java
@@ -3,7 +3,7 @@ package org.fibsters;
 interface DataStorage {
     // Sanitizes and validates the input and output payloads.
     public Result<InputPayload> parseInputPayload(String inputPayloadString);
-    public Result<OutputPayload> parseOutputPayload(ComputeJob CompletedJob);
+    public Result<OutputPayload> parseOutputPayload(ComputeJob completedJob);
 
     // Saving the payload to the database or to a file.
     Result<OutputPayload> saveToDatabase(InputPayload outputPayload);
