+++ src/main/java/org/fibsters/InputPayload.java
@@ -1,0 +1,6 @@
+package org.fibsters;
+
+interface InputPayload extends Payload {
+    public String getUUID();
+    public String getInputType();
+}
